makefile setup for win32 and wasm32

+++ src/common/scene_a.cpp
@@ -1,5 +1,6 @@
 ﻿#include "scene_a.hpp"
 #include "maths.hpp"
+#include "frame_infos.h"
 
 constexpr auto PI = 3.14159265359f;
 constexpr auto HPI = PI / 2.0f;
@@ -10,16 +11,16 @@ void scene_a(void* data, const FrameInfos& frame_infos)
 	constexpr int BSIZE = 50;
 	constexpr int BSIZE_SQUARED = BSIZE * BSIZE;
 
-	const auto h_range = frame_infos.pixel_buffer_width - BSIZE * 2;
-	const auto v_range = frame_infos.pixel_buffer_height - BSIZE * 2;
+	const auto h_range = static_cast<int>(frame_infos.pixel_buffer_width) - BSIZE * 2;
+	const auto v_range = static_cast<int>(frame_infos.pixel_buffer_height) - BSIZE * 2;
 
-	const auto ball_incr = HVEL * frame_infos.time;
+	const auto ball_incr = static_cast<int>(HVEL * frame_infos.time);
 
 	const auto draw_ball_x = abs(
-		(static_cast<int>(ball_incr + h_range) % (h_range * 2)) - h_range) + BSIZE;
+		(ball_incr + h_range) % (h_range * 2) - h_range) + BSIZE;
 
 	const auto draw_ball_y = abs(
-		(static_cast<int>(ball_incr + v_range) % (v_range * 2)) - v_range) + BSIZE;
+		(ball_incr + v_range) % (v_range * 2) - v_range) + BSIZE;
 
 	unsigned* ptr = frame_infos.pixel_buffer;
